# Conflicts:
#	src/main/java/com/ezban/admin/controller/AdminController.java
#	src/main/java/com/ezban/admin/controller/AdminLoginController.java
#	src/main/java/com/ezban/admin/model/AdminSecurityConfig.java
#	src/main/java/com/ezban/host/model/HostSecurityConfig.java
#	src/main/resources/templates/backstage/adminmanage/adminmanage.html
#	src/main/resources/templates/backstage/adminmanage/eventreportback.html
#	src/main/resources/templates/backstage/adminmanage/productreportback.html
#	src/main/resources/templates/frontstage/product/productcomment.html

+++ src/main/java/com/ezban/product/model/ProductDto.java
@@ -51,7 +51,7 @@ public class ProductDto {
     private String productRemoveTime;
 
     @NotNull
-    private Integer productStatus;
+    private Byte productStatus;
 
     public Integer getProductCategoryNo() {
         return productCategoryNo;
@@ -133,11 +133,11 @@ public class ProductDto {
         this.productRemoveTime = productRemoveTime;
     }
 
-    public Integer getProductStatus() {
+    public Byte getProductStatus() {
         return productStatus;
     }
 
-    public void setProductStatus(Integer productStatus) {
+    public void setProductStatus(Byte productStatus) {
         this.productStatus = productStatus;
     }
